requirements.txt
# PTT Scraper Dependencies
selenium-driverless==1.7.0
beautifulsoup4==4.12.2
requests==2.31.0
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
python-dotenv==1.0.0
html5lib==1.1
selenium==4.15.2
webdriver-manager==4.0.1
